Cambio valor a entero.

+++ db/migrate/20211002175138_add_bets_to_bet.rb
@@ -1,7 +1,6 @@
 class AddBetsToBet < ActiveRecord::Migration[6.1]
   def change
-    
-    add_column :bets, :bet, :double
+    add_column :bets, :bet, :int
     add_column :bets, :result, :string
   end
 end
